dh: avoid possible integer underflow using type enforcement.

+++ src/dh.c
@@ -315,7 +315,7 @@ void dh_end_session(void *session)
     MyFree(si);
 }
 
-char *dh_get_s_public(char *buf, int maxlen, void *session)
+char *dh_get_s_public(char *buf, size_t maxlen, void *session)
 {
     struct session_info *si = (struct session_info *) session;
     char *tmp;
@@ -338,7 +338,7 @@ char *dh_get_s_public(char *buf, int maxlen, void *session)
     return buf;
 }
 
-int dh_get_s_shared(unsigned char *buf, int *maxlen, void *session)
+int dh_get_s_shared(unsigned char *buf, size_t *maxlen, void *session)
 {
     struct session_info *si = (struct session_info *) session;
